+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## ????
 - Github Actions for tests
+- Static method, HTTP::g() for GET
+- Static method, HTTP::p() for POST
 
 ## 0.0.3 - 2023-04-10
 - Provide default HTTP options